Ajout de la date dans Restore-TFSBDStorage et /Restore-bd.ps1

+++ scripts/Restore-bd.ps1
@@ -27,7 +27,7 @@ Param(
 
     [Parameter(Mandatory = $false,
         HelpMessage = "Donner la date du restore dans le format yyyyMMdd, sinon on prendra le dernier backup")]
-    $date,
+     $date,
 
     [Parameter()]
     [string]
@@ -41,47 +41,51 @@ if (get-module -ListAvailable AzureRm) {
     if (get-module -ListAvailable Az.sql) { import-module az.sql } 
 }
 
-if (!$date) {
-    $context = Get-AzureRmStorageAccount -StorageAccountName gumbackups -ResourceGroupName infrastructure
-    $dateformat = ((Get-AzureStorageBlob -Context $context.context -container sql-backup -Blob BdGum-prd_* | sort-object -Property LastModified -Descending)[0]).lastmodified
-    $date = "{0:yyyyMMdd}" -f $dateformat
-}
-
-switch ($BD) {
-    "Gum" {
-        $server = "sqlgum-$Destination"
-        $resourcegroup = "Gumsql-rg-$Destination"
-        $BdDest = "BdGum-$destination"
-        $BdSource = "BdGum-$Source_$date"
-    }
-    "AppsInterne" {
-        $server = "sqlguminterne-$Destination"
-        $resourcegroup = "sqlapps-rg-$Destination"
-        $BdDest = "BdAppsInterne-$destination"
-        $BdSource = "BdAppsInterne-$Source_$date"
-    }
-    "Veille" {   
-        $server = "sqlguminterne-$Destination"
-        $resourcegroup = "sqlapps-rg-$Destination"
-        $BdDest = "BdVeille-$destination"
-        $BdSource = "BdVeille-$Source_$date"
-    }
-}
-
-Write-output " Le serveur = $server `n Resourcegroup = $resourcegroup `nBD source = $Bdsource`nBD destination = $BdDest"
-
 [string] $Storagekey = (Get-azureRMStorageAccountKey -ResourceGroupName infrastructure -Name gumbackups ).value[0]
 $StorageAccessKey = [Microsoft.Azure.Commands.Sql.ImportExport.Model.StorageKeyType]::StorageAccessKey
 
 $AdministratorLogin = "sqladmin" + $Destination
 $pass = (Get-azureKeyVaultSecret -VaultName gumkeyvault -name $("sqladmin" + $Destination )).secretvalue
 
-$gum = Get-AzureRmStorageAccount -StorageAccountname gumbackups -resourcegroupname infrastructure
+$Context = (Get-AzureRmStorageAccount -StorageAccountname gumbackups -resourcegroupname infrastructure).context
 
+if ($date) {$date= "{0:####-##-##}" -f $date}
+if (!$date) {
+    $BDs = Get-AzureStorageBlob -Context $context -container sql-backup  | Where-object { $_.Name -like ("Bd" + $BD + "-" + $source + "_*") }
+    if (!$BDs) { Write-Warning "$BD-$source n'existe pas" ; break }
+    if ($BDs -is [array]) { $dateFormat = (($BDs| sort-object -Property LastModified -Descending)[0]).lastmodified }
+    $date = "{0:yyyy-MM-dd}" -f $dateformat
+}
+
+switch ($BD) {
+    "Gum" {
+        $server = "sqlgum-$Destination"
+        $resourcegroup = "Gumsql-rg-$Destination"
+        $BdDest = "BdGum-$Destination"
+        $BdSource = "BdGum-" + $Source + "_" + $date
+    }
+    "AppsInterne" {
+        $server = "sqlguminterne-$Destination"
+        $resourcegroup = "sqlapps-rg-$Destination"
+        $BdDest = "BdAppsInterne-$Destination"
+        $BdSource = "BdAppsInterne-" + $Source + "_" + $date
+    }
+    "Veille" {   
+        $server = "sqlguminterne-$Destination"
+        $resourcegroup = "sqlapps-rg-$Destination"
+        $BdDest = "BdVeille-$Destination"
+        $BdSource = "BdVeille-" + $Source + "_" + $date
+    }
+}
+
+Write-output "Le serveur = $server `nResourcegroup = $resourcegroup `nBD source = $Bdsource`nBD destination = $BdDest"
+
+$Bdname = Get-AzureStorageBlob -Context $context -Container sql-backup | Where-Object { $_.name -like ($bdSource + "*") } ;
+if ( !$Bdname ) { write-warning "$BdSource n'existe pas"; break} 
+if ($Bdname -is [array]) { $Bdname = ($Bdname | Sort-Object -Descending LastModified )[0]}
 $databases = get-azureRMsqlserver -name $server -resourcegroupname $resourcegroup | get-azureRMsqldatabase | where-object { $BdDest -contains $_.Databasename } 
-$databases | Remove-AzureRMsqldatabase 
+if ( $databases) { $databases | Remove-AzureRMsqldatabase }
 
-$Bdname = (Get-AzureStorageBlob -Context $gum.context -Container sql-backup | Where-Object { $_.name -like ($bdSource + "*") } | Sort-Object -Descending LastModified  )[0];
 $Restore = New-azureRMSqlDatabaseImport `
     -ServerName $server `
     -DatabaseName $BdDest `
@@ -95,6 +99,4 @@ $Restore = New-azureRMSqlDatabaseImport `
     -ServiceObjectiveName S0 `
     -DatabaseMaxSizeBytes 30gb ;
 
-$restore
-
-
+return $restore.OperationStatusLink
